refactor(ai): extract useAiConfig and useAiUsage hooks from page.tsx
Extract AI config state management (config, savedConfig, loading, saving,
hasUnsavedChanges, loadData, handleSave) into useAiConfig hook.

Extract AI usage state management (usageSummary, todaySpend, loadingUsage,
loadUsageData) into useAiUsage hook, used by TabConsumo.

Both hooks are in configuracoes/ai/hooks/ with barrel export.

Fixes #244
